backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
fastapi-admin==1.0.3
bcrypt==4.1.2
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.9 